🎨 made function (translate vaccine's name)
