+++ src/pages/mypageResInquiry.jsx
@@ -6,6 +6,7 @@ import MyPageBoxSet from "../components/mypage/mypageBoxes";
 import MyPageBtn from "../components/mypage/mypageBtn";
 import axios from "axios";
 import { useState, useEffect } from "react";
+import transVaccine from "../components/modules/translation";
 const MyPageResInquiry = () => {
   const [ssn, setSsn] = useState("");
   const [name, setName] = useState("");
@@ -26,13 +27,6 @@ const MyPageResInquiry = () => {
         setResInfo(response.data);
       })
       .catch((e) => console.log(e));
-  };
-
-  const transVaccine = (name) => {
-    if (name == "Pfizer") return "화이자";
-    else if (name == "Morderna") return "모더나";
-    else if (name == "AstraZeneca") return "아스트라제네카";
-    else if (name == "Janssen") return "얀센";
   };
 
   const transDate = (date) => {
